store_data_in_cloud_storage/requirements.txt
google.cloud==0.34.0
google-cloud-storage==2.10.0